+++ src/select/services.js
@@ -78,14 +78,24 @@ angular.module('oi.select')
      *
      * @param {DOM element} container
      * @param {DOM element} contained
+     * @param {string} class name of element in container
      * @returns {boolean}
      */
-    function contains(container, contained) {
+    function contains(container, contained, className) {
         var current = contained;
 
         while (current && current.ownerDocument && current.nodeType !== 11) {
-            if (current === container) {
-                return true;
+            if (className) {
+                if (current === container) {
+                    return false;
+                }
+                if (current.classList.contains(className)) {
+                    return true;
+                }
+            } else {
+                if (current === container) {
+                    return true;
+                }
             }
             current = current.parentNode;
         }
@@ -108,7 +118,7 @@ angular.module('oi.select')
         inputElement.on('focus', focusHandler);
 
         function blurHandler(event) {
-            var relatedTarget = event.relatedTarget;
+            var relatedTarget = event.relatedTarget; //TODO: get relativeTarget in IE, FF (event.explicitOriginalTarget || document.activeElement);
 
             if (relatedTarget === inputElement[0]) {
                 event.stopImmediatePropagation(); //cancel blur if focus to input element